refactor(search applications): rename generate api key modal files
