+++ x-pack/plugins/enterprise_search/public/applications/applications/components/search_application/connect/generate_api_key_modal/generate_search_application_api_key_modal.tsx
@@ -33,7 +33,7 @@ import { i18n } from '@kbn/i18n';
 
 import { GenerateEngineApiKeyLogic } from '../../../../../enterprise_search_content/api/generate_engine_api_key/generate_engine_api_key_logic';
 
-import { GenerateApiKeyModalLogic } from './generate_engine_api_key_modal.logic';
+import { GenerateApiKeyModalLogic } from './generate_search_application_api_key_modal.logic';
 
 interface GenerateEngineApiKeyModalProps {
   onClose(): void;
